Wire full pipeline and implement all viz components
app/main.py: runs the complete pipeline on button click (DatasetProcessor
→ EmbeddingPipeline → classifier → simulate_batches → per-batch geometry
+ eval) and caches results in session_state. Viz components now receive
pre-computed BatchResult lists instead of raw data.

app/utils.py: adds BatchResult dataclass holding all per-batch outputs.
upload.py: returns (df, label_col) using DatasetProcessor.get_label_columns.
embedding_viz.py: 2D PCA scatter with batch slider + drift score trend.
cluster_viz.py: centroid shift heatmap, reassignment rate line, per-batch bar.
retrieval_viz.py: neighbor instability + distance shift line charts + table.
performance_viz.py: dual-axis metric/drift chart + calibration curve.

Fixes: calibration_curve import (sklearn.calibration in sklearn>=1.x);
ClusterTracker.fit/predict cast inputs to float64 for sklearn KMeans.

+++ analysis/cluster_tracker.py
@@ -17,13 +17,13 @@ class ClusterTracker:
     def fit(self, baseline_emb: np.ndarray) -> "ClusterTracker":
         """Fit k-means on baseline embeddings. Frozen after this call."""
         self._kmeans = KMeans(n_clusters=self.n_clusters, random_state=self.random_state, n_init="auto")
-        self.baseline_labels_ = self._kmeans.fit_predict(baseline_emb)
+        self.baseline_labels_ = self._kmeans.fit_predict(baseline_emb.astype(np.float64))
         self.baseline_centroids_ = self._kmeans.cluster_centers_.copy()
         return self
 
     def predict(self, emb: np.ndarray) -> np.ndarray:
         """Assign embeddings to baseline clusters (no refit)."""
-        return self._kmeans.predict(emb)
+        return self._kmeans.predict(emb.astype(np.float64))
 
     def centroid_shift(self, emb: np.ndarray) -> np.ndarray:
         """Mean distance each cluster centroid has moved (using current batch embeddings).
